+++ voxygen/src/hud/mod.rs
@@ -645,6 +645,34 @@ pub struct MapMarkers {
     group: HashMap<Uid, Vec2<i32>>,
 }
 
+/// (target slot, input value, inventory quantity, is our inventory, error,
+/// trade.offers index of trade slot)
+pub struct TradeAmountInput {
+    slot: InvSlotId,
+    input: String,
+    inv: u32,
+    ours: bool,
+    err: Option<String>,
+    who: usize,
+    input_painted: bool,
+    submit_action: Option<TradeAction>,
+}
+
+impl TradeAmountInput {
+    pub fn new(slot: InvSlotId, input: String, inv: u32, ours: bool, who: usize) -> Self {
+        Self {
+            slot,
+            input,
+            inv,
+            ours,
+            who,
+            err: None,
+            input_painted: false,
+            submit_action: None,
+        }
+    }
+}
+
 pub struct Show {
     ui: bool,
     intro: bool,
@@ -676,6 +704,7 @@ pub struct Show {
     prompt_dialog: Option<PromptDialogSettings>,
     location_markers: MapMarkers,
     salvage: bool,
+    trade_amount_input_key: Option<TradeAmountInput>,
 }
 impl Show {
     fn bag(&mut self, open: bool) {
@@ -1089,6 +1118,7 @@ impl Hud {
                 prompt_dialog: None,
                 location_markers: MapMarkers::default(),
                 salvage: false,
+                trade_amount_input_key: None,
             },
             to_focus: None,
             //never_show: false,
@@ -2811,7 +2841,7 @@ impl Hud {
         }
         // Trade window
         if self.show.trade {
-            match Trade::new(
+            if let Some(action) = Trade::new(
                 client,
                 &self.imgs,
                 &self.item_imgs,
@@ -2822,23 +2852,36 @@ impl Hud {
                 i18n,
                 &msm,
                 self.pulse,
+                &mut self.show,
             )
             .set(self.ids.trade, ui_widgets)
             {
-                Some(action) => {
-                    if let TradeAction::Decline = action {
-                        self.show.stats = false;
-                        self.show.trade(false);
-                        if !self.show.social {
-                            self.show.want_grab = true;
-                            self.force_ungrab = false;
-                        } else {
-                            self.force_ungrab = true
-                        };
-                    }
-                    events.push(Event::TradeAction(action));
-                },
-                None => {},
+                match action {
+                    Err(update) => match update {
+                        trade::HudUpdate::Focus(idx) => self.to_focus = Some(Some(idx)),
+                        trade::HudUpdate::Submit => {
+                            let key = self.show.trade_amount_input_key.take();
+                            key.map(|k| {
+                                k.submit_action.map(|action| {
+                                    self.events.push(Event::TradeAction(action));
+                                });
+                            });
+                        },
+                    },
+                    Ok(action) => {
+                        if let TradeAction::Decline = action {
+                            self.show.stats = false;
+                            self.show.trade(false);
+                            if !self.show.social {
+                                self.show.want_grab = true;
+                                self.force_ungrab = false;
+                            } else {
+                                self.force_ungrab = true
+                            };
+                        }
+                        events.push(Event::TradeAction(action));
+                    },
+                }
             }
         }
 
@@ -3859,15 +3902,15 @@ impl Hud {
         scale_mode
     }
 
-    // Checks if a TextEdit widget has the keyboard captured.
-    fn typing(&self) -> bool {
-        if let Some(id) = self.ui.widget_capturing_keyboard() {
-            self.ui
-                .widget_graph()
+    /// Checks if a TextEdit widget has the keyboard captured.
+    fn typing(&self) -> bool { Hud::is_captured::<widget::TextEdit>(&self.ui.ui) }
+
+    /// Checks if a widget of type `W` has captured the keyboard
+    fn is_captured<W: Widget>(ui: &conrod_core::Ui) -> bool {
+        if let Some(id) = ui.global_input().current.widget_capturing_keyboard {
+            ui.widget_graph()
                 .widget(id)
-                .filter(|c| {
-                    c.type_id == std::any::TypeId::of::<<widget::TextEdit as Widget>::State>()
-                })
+                .filter(|c| c.type_id == std::any::TypeId::of::<<W as Widget>::State>())
                 .is_some()
         } else {
             false